+++ .eslintrc
@@ -1,4 +1,5 @@
 {
+  "parser": "babel-eslint",
   "extends": "airbnb/base",
   "rules": {
     "semi": [2, "never"],
