+++ lib/gengi.js
@@ -1,7 +1,7 @@
-const fetch = require('node-fetch')
-const chalk = require('chalk')
+import fetch from 'node-fetch'
+import chalk from 'chalk'
 
-module.exports = (currency, value) => {
+export default (currency, value) => {
   fetch(`http://api.gengi.is/calculate/${currency}/${value}`).then(response => {
     if (response.status !== 200) {
       const error = new Error(response.statusText)
