+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "gengi-cli",
-  "version": "1.0.3",
+  "version": "1.0.4",
   "description": "",
   "main": "index.js",
   "scripts": {
@@ -10,6 +10,9 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "babel-plugin-add-module-exports": "^0.1.2",
+    "babel-preset-es2015": "^6.6.0",
+    "babel-register": "^6.7.2",
     "chalk": "^1.1.1",
     "node-fetch": "^1.3.3"
   },
@@ -17,6 +20,7 @@
     "gengi": "./bin/gengi"
   },
   "devDependencies": {
+    "babel-eslint": "^6.0.0",
     "eslint": "^2.5.1",
     "eslint-config-airbnb": "^6.2.0"
   }
